- add module-alias back. - refactor and add a new cli folder.
