+++ src/cli/commands/config/schema.ts
@@ -1,0 +1,9 @@
+import * as yup from 'yup';
+
+export const schema = yup
+  .object({
+    path: yup.string().required(),
+  })
+  .required();
+
+export type Schema = yup.InferType<typeof schema>;
