+++ src/cli/commands/index.ts
@@ -1,0 +1,2 @@
+export { action as start } from './start/action';
+export { action as config } from './config/action';
